pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finsde"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "matplotlib",
    "numpy",
    "pandas",
    "scikit-learn",
    "hydra-core",
    "wandb",
    "torch",
    "lightning",
]

[tool.setuptools.packages.find]
include=["finsde"]